[libringbuffer] fix split buffer read/write

+++ gear-lib/libringbuffer/libringbuffer.h
@@ -39,10 +39,10 @@ struct ringbuffer *rb_create(int len);
 void rb_destroy(struct ringbuffer *rb);
 ssize_t rb_write(struct ringbuffer *rb, const void *buf, size_t len);
 ssize_t rb_read(struct ringbuffer *rb, void *buf, size_t len);
-void *rb_dump(struct ringbuffer *rb, int *len);
+void *rb_dump(struct ringbuffer *rb, size_t *len);
 void rb_cleanup(struct ringbuffer *rb);
-int rb_get_space_free(struct ringbuffer *rb);
-int rb_get_space_used(struct ringbuffer *rb);
+size_t rb_get_space_free(struct ringbuffer *rb);
+size_t rb_get_space_used(struct ringbuffer *rb);
 
 #ifdef __cplusplus
 }
